Make Piece a struct
The term "piece" is often used to refer to the _type_ of a piece. That
is, whether it is a Knight, Queen, etc. However, in general, a piece is
meant to also ecapsulate a color as well. That is, a piece is both a
type and a color.

Type is a very generic word, so use Rank as the name of the enum for the
types of pieces. Then make Piece a struct that has rank and color
members.

This will make it easier to construct the board, because the squares on
the board will simply be either empty, or taken by a piece.