Make Ply a trait
The primary reason for making Ply a trait is so that it's variants,
which are now independant structs that implement Ply, can have a
validate method that properly knows how to validate a move given a
board.

Without using a trait, a single method, perhaps then split into other
methods using a match, would have to be used, creating a somewhat
tangled mess of code around this simple concept. Instead, this allows
us to delegate that logic to the implementation of the different Ply
types.

One unfortunate side-effect is that the ply types are no longer
namespaced. While not a significant concern, I think that it is a bit
of an ergonomic loss. Hopefully I'll be able to find a reasonable
alternative in the near future.

They are also not matchable at all. This may prove to be a significant
hurdle in the future when writing move parsing code, but for now, it
creates a very simple implementation. Overcoming it may require
creating another enum with variants that wrap these types, and thereby
make the options matchable.

The tests also had to be rewritten to take the new types into account.