Revert "Make Ply a trait"
This reverts commit 88da2ddee0ef58258138d05af9c852209c95fbe6.

After discussion on the #rust IRC channel, the general consensus is that
using a trait to do OO-style method overriding when an enum is likely
more obvious is a bad idea. Better to live with the big match than to be
sucked into using traits when you don't need to.