Create main binary
In addition to being a solid library, the chess package should also end
up as a decent command line program with a playable game. This is
possible because a package can be a library as well as a binary.

Because the main function isn't tested, and may be difficult to test,
using submodules is a pain, because they show up as unused imports. The
main function as well shows up as an unused function when testing. Since
main is such a special case, there is an open issue to make the test
runner not show that warning.

https://github.com/rust-lang/rust/issues/12327

For the time being though, I've done as the OP of that issue said he
does: putting #[cfg(not(test))] before the main function so that the
linter ignores it.