Rename Notation trait to PlyInputNotation
The reason for the rename is to allow particular notations to be used
for varying purposes. For example, the ZeroIntegersNotation could be
extended to understand how to unparse a board into a string
representation. Since not every kind of notation will be able to do all
of the things that may be desired of any notation, it makes sense to
split the different uses into separate traits. We may wish to re-address
the split of the traits at some point in the future, after we've seen
how the current plan expands.