Add DisplayBoardNotation struct
It implements BoardOutputNotation so that it can print. However, it also
delegates some functions to a BoardOutputNotationHelper trait so that
the logic for displaying the text-based board is separated from the
logic to show the pieces and the file and rank labels.

Notably incomplete are the stubbed implementations of unparse_castling
and unparse_enpassant of the BoardOutputNotationHelper trait. They are
stubbed to make the opening board test correctly.